+++ RELEASE_NOTES.md
@@ -4,6 +4,7 @@
 
 - introduce console with basic operations
 - Add meta-collection collection to store metadata about collections and indices
+- Add multi collection type to support collection with several values for the same key
 
 ## 2.7 - 2026-04-12